[cm][routing] Stateless routing
This change makes the routing code stateless. What this allows
is for routing to begin at any point in the process. Routing can
start at an offer, or an expose.

This opens the door for multi-step routing, where a capability is
routed to some intermediate declaration, where further action may
cause it route to completion. An example of this is Unified Service
routing.

Change-Id: Ic1f003aa2e1a0cb82db17fa9ac185774e730179e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486383
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Gary Bressler <geb@google.com>