[stdcompat] Add cpp20::invoke and related traits
This adds a conforming implementation of std::invoke as defined in
[func.invoke] and [func.require] ¶ 1 and 2, and the following traits
that also depend on INVOKE:

cpp17::invoke_result
cpp17::is_invocable
cpp17::is_invocable_r
cpp17::is_nothrow_invocable
cpp17::is_nothrow_invocable_r

And their related typedefs and variable templates.

This CL is structured so that the function template declarations
for the 7 points in [func.require] ¶ 1 are available to type_traits.h
so that the traits can be implemented without needing to include
functional.h.

We define invoke() in cpp20:: instead of cpp17:: (the standard when it
was introduced) because our version is constexpr, which is in accord
with the C++20 standard but not C++17. However, the internals are
identical to the C++17 version, so the internal namespace is still
cpp17.

Test: stdcompat-tests
Change-Id: I386b14b725f541e49e693b726795cf77bbbb5ef5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474405
Commit-Queue: Mark Schott <schottm@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>