[kernel][vm] Use gen count to avoid re-querying pages
Typically it is expected that a copy needs to be re-tried due to having
to fault in a mapping for a different vmo, and that this VMO is not
being actively modified in parallel. Using the generation count allows
detecting this common case and immediately re-trying to copy without
having to re-walk the VMO hierarchy and page list.

Change-Id: Ieae0cad0b0333f610c32c3085392cf55e6df8f77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493219
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>