[timekeeper] More resiliance on RTC time parsing.
From fxb/70974 it looks like the RTC sent an invalid time (e.g. month=0,
day=0). Which led to a panic in timekeeper. Thats undesirable since we
see an RTC as only a "nice to have". This handles the error gracefully
and includes the invalid time in the error message so we should be able
to isolate the problem if it happens again.

Bug: 70974
Change-Id: I7f5b81bdedc9ee3ebc2fad99ad027a44ccc3f277
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493660
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>