[fidl][fuzzing] Seatbelt for cpp FIDL bindings fuzzer
The fuzzer was reading a uint64_t and using it to bound the size of a
vector of handles. Naturally, this allowed for instantiating more
handles than the allocator is willing to allocate. This change bounds
the vector at 2 * max number of handles in a FIDL message (to test an
oversized but reasonable collection of handles).

Bug: 71159
Test: fx fuzz repro fidl-fuzzers/conformance-fuzzer [testcase-file-linked-in-bug]
Change-Id: I638665bf971917aba015c4aa8eaa3e2955d051fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494238
Commit-Queue: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>