[time] Verify time source is kept alive if a dependency is unhealthy
Adds a test that verifies that an inactive time source is not restarted
if it has declared that some dependency is unhealthy.

This also refactors the PushSourcePuppet and allows it to track the
number of connections it has received. Tests that previously relied on
Cobalt events to verify the restart of a time source now inspect this
number instead.

Bug: 61500

Multiply: timekeeper-integration

Change-Id: If3a0e407495722e57fc444ee888a3dd349b87326
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493664
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>