[kernel][x86] Migrate swapgs bug mitigation to v2 patching
This is the first migration of the v1 code-patching over to the v2. As
mentioned in the bug, we will migrate over to v2 within the kernel ahead
of physboot's schedule. The pre-vm init hook that currently does the v1
code-patching now does both. We migrate the swapgs bug patching first as
it is one of the more simpler cases.

The v2 patching should replace the three one-byte `nop`s of v1 with a
single three-byte `nop`.

Successfully boots on both an Intel NUC and a AMD Zen.

Bug: 67615
Change-Id: I72e608b3b8e257c080ef87099e1cff86b54f7415
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483180
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>