[settings] flush input writes immediately
If the user hard reboots the device shortly after changing the sw mute,
the pending changes don't have a chance to flush and are lost. This
changes to flush immediately so that they are not lost in this case.

Follow-up bug for considering other settings: 71292

Bug: b/181571540
Test: fx test setui_service_test

Change-Id: Ia90b2d153a40515b5c8efb052ec6479bbf6e159d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493987
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>