[docs] Add bringup product definition
This defines a most minimal target for development. Note that testing
was purposely left out of this definition (as in what tests are run via
`runtests` versus `fx test` and available in which product).
Note also that more advanced product configurations with improved
workflows for example via overnet over serial are not prevented by
this definition.

Change-Id: I2f64b851b41ef203e9dc6b2caee7486b6508d0d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492640
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Curtis Galloway <curtisgalloway@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>