[kernel][lockup_detector] Fix a gap in CS coverage
Recent changes to the critical section lockup detector made it so that
threads exiting a critical section were not responsible for reporting
long running critical sections. Instead, heartbeat check-ins became
responsible for noticing that a thread had started, but not ended a
critical section for a long time (default, 3 sec). This is good
because it allows us to detect permanent critical section hangs before
the WDT has time to fire.

Unfortunately, it means that we lost the ability to notice long, but
not fatal amounts of time spent in critical sections (such as SMC
calls).

Restore some of this capability by introducing the per-cpu concept of
"worst case critical section time".  Threads exiting a CS now check to
see if the time spent in the CS was worse than their worst time, but
they do not make any effort to report it.  Instead, they just update a
per-cpu value, and let another checker report the value.

Checkers will only report a worst case value (as an OOPS) when it
1) Exceeds the minimum kcounter bucket threshold
2) Is worse than the last reported worst case
3) The rate limiter will permit it.

Additionally, the worst case for each core can be queried using
`k lockup status`

Bug: 70724
Bug: 69037
Change-Id: I1f1cfec30bd76ab69554130b57515afda7bdda91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494721
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>