[debugger] Fix the unwinder failure for abort()
libunwindstack expects a pc adjustment in order to unwind correctly. We
used to disable this feature because we want the return addresses
instead of call sites, which caused the failed unwinding for abort().

This patch enables the pc adjustment in libunwindstack but also restores
the pc after libunwindstack returns, so that the unwinding is improved
but the original behavior is preserved.

There's still an issue with the display of the frames when the pc is
outside of a function, which could happen when the function is calling a
noreturn function as its last instruction.

Fixed: 45306, 58504
Change-Id: I2fec31155e4a9c7ed0e68897541b0471125f1c3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493001
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>