[kernel][lockup] Add new cmdline params and update docs.
Add two new kernel command line parameters which control the "fatal"
thresholds for the critical section and heartbeat checkers.

Soon, if these thresholds are enabled and exceeded, it will trigger
generation of a crashlog with a reboot reason of SOFTWARE_WATCHDOG,
and reboot the system.  Note that there is no actual change in
behavior, yet.  These parameters are being introduced a bit early in
order to get out of the way of the boot-options train, and to make it
easier to pick these changes into the F1 branch (which will not have
boot-options support)

Bug: 70724
Change-Id: Ifb78f20225bf4a2828bd0a668b997a6f574bf755
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495060
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>