[go][gidl][fidl] Handle rights GIDL tests for Go
This CL is used in conjunction with:
https://fuchsia-review.googlesource.com/c/third_party/go/+/493283

This enables support for checking handle rights and
types in the Go GIDL tests. More or less all that is
needed for this from the fuchsia.git side is
populating handle rights information on structs used
for checking.

Bug: 41920
Change-Id: Ibdbb5839c6263967ef50c0b6d720527716fbd9cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493599
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>