[docs][wdt] Clarify some confusing wording.
Clarify some of the wording in the WDT theory of operation doc.
Specifically "may not" (indicating that something is not permitted to
happen) was used when "might not" (meaning that it may or may not
happen) was the intent.

Also fix a mismatch where a second person pronoun (You) was paired
with a third person pronoun (They) later in the sentence.  Address
this by eliminating the You, and replacing it with a more appropriate
specific noun (Developers).

Change-Id: I45ce334c53a3b1f6f9dbbc10c478c3fb04a34228
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494498
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>