[fidlc] Pipe syntax flag through to Consume step
This CL sets the groundwork for switching behavior based on a
per file syntax. The current code uses a global experimental
flag in such a way that it can be mechanically replaced with
a per file version boolean when that is supported in the syntax.
The comment on ConsumeStep in flat_ast.h explains how this
value flows through the compiler.

To actually make use of this, support for protocols, services,
and resource declarations are added for in the new syntax. This
code path will eventually be tested once we update existing
tests to run on both the old and new syntax.

Change-Id: I4f894e248521ca41c3aa89dd5be4c8ced6d03a5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492398
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Commit-Queue: Felix Zhu <fcz@google.com>