[appmgr] Set up and test directory walker for component storage.
This is the first change with associated tests to set up per-component
storage metrics export. The follow on changes will be to push them to
inspect, convert the component path names to useful monikers, and get
the data out to cobalt.

Currently this takes a list of directories to sum and periodically does
so via the main dispatcher.

Bug: 69017

Change-Id: I29668827086ab81fb375804464518322a6d72ae5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491339
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Vardhan Mudunuru <vardhan@google.com>