[hlcpp] Handle rights support for HLCPP conformance tests
This enables handle rights and object type checking in GIDL-generated
conformance tests for HLCPP.

There currently isn't a mechanism to check handle rights and type after
decoding and the handle id changes if there are too many handles
because the handle is replaced. Because of this, handle decode tests
are temporarily disabled with a TODO. I plan to revisit this and build
functionality for rights checking, but I hope to make incremental
progress by getting the rest of the conformance test handle rights
functionality working.

Bug: 41920
Change-Id: Ib494d24529ef6e827a9e33a458bb2ce9a97bf98c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495178
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>