[fdio] Simplify ppoll and select
Use deferred cleanup to avoid various gymnastics.

This also fixes the behavior of both functions when zero fds are passed
and the timeout is a null pointer; this change is untested because the
correct behavior in that case is to block indefinitely.

This also correctly returns an error from ppoll when a negative timeout
is passed.

Change-Id: Ic6150e24cf0092cf107dbbf182d1bb5eaa51fd0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495239
Commit-Queue: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Mithun Iyer <iyerm@google.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>