[wlan] Add MLME event that triggered disconnect to Inspect tree
The wlanstack.cmx Inspect tree no longer includes the
disconnect_reason field and has an expanded disconnect_source field.

The new format of the disconnect_source field is following:

  "source: <user|ap|mlme>, reason: <disconnect reason>"

If the source if ap or mlme, then the string will have the following
appended:

  ", mlme_event_name: <DisassociateIndication|DeauthenticateIndication>"

Including the mlme event that triggered the disconnect adds an extra
dimension which can be used when determine the type of event which
triggered a disconnect.

Note: Some F1 specific changes made to tie-over the lack of the new
fuchsia.wlan.ieee80211 ReasonCode type not being refactored through
in the F1 branch yet.

Test: fx test wlanstack-tests wlan-sme-tests
Bug: 70687
Change-Id: I15ec6733f5121692f1804edde9958fc161d51bdd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496022
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>