[kernel][lockup][ddk] Downgrade SMC overrun from oops to warning
We're seeing a fair number of these in the 10-100ms range.  Prior to
this change an overrun of 10+ ms would trigger a kernel OOPS
which (via tefmocheck) will cause tests running at the time to be
classified as failures.

Downgrade the OOPS to a warning and remove the message-specific
tefmocheck so that we don't unnecessarily fail tests and waste infra
resources.

Also, decrease the rate limiter period from 10s to 1s so that we'll
get better visibility into repeated overruns that exceed the
threshold.

Bug: 71167
Change-Id: If41c96ec765e4f9eef4f31c9aac842d2a698843e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495899
Reviewed-by: RJ Ascani <rjascani@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>