[tracer] Fix for nonexisting depfile
Action tracer expects the depfile to exists, which
may not be true in many cases, so it starts complaining
of invalid writes when creating depfiles. So fix the
action tracer such that, even if it doesn't exist,
allow writing the depfile.

Bug: 71561
Test: fx build/tracer

Change-Id: I347fd1c9a8eedfe5b0592190da8af86a91e535f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496101
Commit-Queue: Sundarajan Srinivasan <sundarsrini@google.com>
Fuchsia-Auto-Submit: Sundarajan Srinivasan <sundarsrini@google.com>
Reviewed-by: David Fang <fangism@google.com>