[scenic] Fixed image creation bug.
Since image creation commands can be enqueued on a Scenic session
without immediately presenting, this can create a large enough
window for the buffer collection that the image is created on to be
deregistered in the meantime.

The buffer collection, however, doesn't immediately get freed, which
means that it can still be used for the late image creation. This
patch adds a testcase for this issue and adds a fix for it.

fx test scenic_tests

Change-Id: I9cea8a3a1364076942416ba3c1d822d2d488c3b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495960
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: David Reveman <reveman@google.com>