[kernel][vm] Replace PageQueues Spinlock with CriticalMutex
Most page queue operations that acquire the PageQueues::lock_ are
expected to complete quickly. The exceptions are GetPagerQueueCounts()
which is used to query pager queue lengths, and debug functions used in
tests and k commands, which perform similar entire queue level
operations.

Holding a spinlock across these long running operations can leave
interrupts disabled for several milliseconds. Use CriticalMutex (mutex
which disables preemption) instead of a spinlock, so that interrupts are
not delayed.

CriticalMutex is chosen over a regular Mutex to avoid cases where the
thread holding the lock gets preempted, delaying several other waiters.
The PageQueues lock is a critical lock in the system and is used for
most VM operations.

Bug: 71369
Change-Id: If185a14b673f3f28f71b0185502fb02f605fae25