[recovery-diagnostics] Add proof of concept to check filesystem health
Change-Id: Ib373e3d2f30e3ea9d235077db91752fa7a91d81e

[recovery] Add block listing to diagnostic checks

Also adding additional error handling and robustness in the case
unexpected parts fail.

Change-Id: I52746ef61682078aacf7edb1f58fa1d5884f37f1

[recovery-diagnostics] Wait for file system to bind

It seems release builds often bind the ramdisk before the sys partition.
In this case the diagnostics just need to wait an extra few hundred
milliseconds. Adding a 5 seconds sleep to be extra generous.

Change-Id: Ib7ee584ab39859f179080038aac8b5250898b2b7

[recovery-diagnostics] Read blobfs entries

Reading all of the entries in the root blobfs and trying to hash their
data in case it indicates bad data in the returned file.

Change-Id: I989070db6e36ebeb59c6af4370171606d460a94d