[blobfs] Simplify corruption handler.
Makes the currption handler an abstract class and derive a test version
in the unit tests to verify the corruption handler is called at the
correct times. Simplify the implementation to reduce boilerplate.

The blobfs-integration test is renamed (was misspelled "blobs") which I
noticed while testing this.

There should be no behavior change.

Change-Id: Id02543f18268facdee78b97ebf22fd9c1f5ce967
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495818
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>