[gigaboot] add LZ4 decompression capabilities
Adds functionality to decompress LZ4 images. This will be necessary for
FVM flashing, since the images are large enough that decompression saves
substantial transfer time.

LZ4 decompression must be done in chunks in order to allow time to
service other interfaces - in particular, the network. This
decompression will take long enough that if we did it all in one go, we
would fail to respond to ICMP and/or fastboot packets for too long and
the connection would drop.

This CL adds the device-side support and host-side tests, but nothing is
actually using it yet. Follow-up CLs will hook it into the fastboot
flashing routines.

Bug: 68692
Change-Id: Iab539e27b1b9c8a8c8eb7864004f275d84adfc64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495995
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>