[brcmfmac] Clean up SAE state on authentication failure.
Before this change, the SAE auth state only be cleaned when SAE succeed
while it should be cleaned in both success and failure cases, this
change fixes it.

Also, this change also issues a assoc_mgr_cmd to the firmware to get it
out of waiting state(though an error is returned).

Bug: 71233
Test:
- All current SIM auth tests for WPA3 passed.
- Passed wlan-smoke-test.
- ACTS tests passed(ran by Hayden, with new SL4F test case added for
this case specifically).

Change-Id: I9e17c274708288894df9a74ec941ca1972f7491a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492182
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>