[bind] Decode new bytecode into a bind program
Introduce a DecodedProgram struct that reads in
bytecode and create a symbol table and list of
instructions from it.

When the DecodedProgram receives the bytecode,
it verifies the headers and section. After that,
it splits the sections into separate vectors and
decode them seperatly.

This CL also moves common enums and constants
into a separate file.

Bug: 69608
Test: Added unit tests in decoded_bind_program.rs
Change-Id: Ic41e975f1e93c570882a13a1c6276bc471bdc75e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494819
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>