[kernel][lockup] Reboot when fatal thresholds are exceeded.
Create a crashlog and reboot the system with a reboot reason of "SW
WATCHDOG" when any of the lockup conditions' configured fatal
thresholds are exceeded.

Also, alter the heartbeat check-in behavior to pet the watchdog iff
the heartbeat condition has a configured fatal threshold.  With this
change, we should never see a HW WDT firing as long as there are
active cores who are still checking in on each other.

With these changes, we should be in a position to start collecting
data from production field devices which are hitting the HW WDT.  Now,
instead of hitting the HW WDT, they should end up tripping the SW
WDT (lockup detector) level instead, generating a crashlog containing
details for us to analyze.

Bug: 70724
Change-Id: I0330286dce7aa6323bf45ed699f904b35e9dc139
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496819
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Commit-Queue: John Grossman <johngro@google.com>