[kernel][arm][mmu] make sure kernel page tables are using the global ASID
Switch the kernel TTBR1 register to use the proper global ASID (all 1s)
to match the supposed use. Currently is set to 0 in start.S, which is an
unused ASID. Current behavior is mostly harmless since we dont use the
unused ASID for anything else and we invalidate all kernel mappings with
a global TLB shootdown. However, future fixes in this area will require
the global ASID to be properly carved out as it is.

Reassign GLOBAL_ASID from 0xffff to 1 for future enhancment to handle 8
bit ASIDs properly.

Add an additional safely DSB after turning off the trampoline in start.S

No functional change in this commit.

Bug: 71224
Change-Id: Ia85ede2cae641d18096a5090dbe4e53225006f5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493681
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>