[fidl][go] Part 2/2: Move conformance test file listing to Go repo
This is meant to be combined with:
Idb87484f2ea8e672ffd2fc940225aedaa834f546

Currently, it isn't clear that you need to update a file list to get the
go conformance tests to pass. Additionally, you need to do a multistep
migration to add a file because you need to first add the file to the go
repo, then add the filename to fuchsia.git, then use the file in the go
repo.

Change-Id: I9adc885c64b741f8b9085a75000aeb847c313c40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496140
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>