[scrutiny] Improve the shell to support buffered output.
This change introduces buffered output to the shell which allows dynamic
switching between stdout and a buffered string. This is really nice for
programming against Scrutiny in Rust. Instead of having to deal with
piping stdout you can just get the result directly out of the execute.

Change-Id: Ib7761714580502e751222e3294bcde8459fcdafd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478364
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>