[fidl][test] Add missing dart dependency
Several tests under dart_bindings_test depend on
`test/test/server.dart`, while it's not explicitly listed as a source
to `bindings-test-component`. This mean if `test/test/server.dart` is
changed, it won't correctly trigger a rebuild of this test component
when building incrementally.

To fix it, I wrapped `server.dart` in a library and set it as a
dependency of the test component.

NOTE: At first glance a simpler fix would be to add `server.dart` to
`sources` of the test component directly. However that didn't work
because `dart_test_component` expects a `main` function in all of its
sources and uses that to generate a `main.dart` for test.

Change-Id: I7db2539443b939821116b04875408d0397bdb4a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497485
Reviewed-by: Felix Zhu <fcz@google.com>
Commit-Queue: Jay Zhuang <jayzhuang@google.com>