[media][audio_core] Prevent SetReferenceClock race
In both AudioCapturer and AudioRenderer, a boolean ensures the reference
clock is only set once. The boolean is not protected by a mutex, so a
race can occur when concurrent calls to SetReferenceClock are made. This
CL protects |reference_clock_is_set_| with a mutex.

Test: MULTIPLY: audio_core_unittests
Test: MULTIPLY: audio-core-api-tests
Test: MULTIPLY: audio-core-api-pipeline-tests

Change-Id: Iff787738f1609bc877efbbe29c2438f8c0590692
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483340
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Payton Bradsky <pbradsky@google.com>