[epoch] refactor epoch (de)serialization into a lib
In this change, we refactor epoch.json (de)serialization into a lib.

Justification: Although the underlying implementation is fairly straightforward, we extract this into a library so that the library serves as the "source of truth" for all valid epoch.json formats. Currently, there is only one format documented in [RFC-0071](https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0071_ota_backstop). If we add a new format (e.g. change the version field) in the future, this library will ensure that all clients are aligned on format changes.

Test: fx test -o epoch-tests
Test: fx test -o epoch-generator-tests
Bug: 68406
Change-Id: I9efd51b4bc71e60ac9848d06a7e6dd1596ab8c37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497660
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>