[blobfs] Move code out of blob files.
Moves Blob::WriteInfo out of the header and into the .cc file which
makes the Blob header a bit easier to follow. This struct only needs to
be forward-declared and is used only in blob.cc.

Moves the "Producer" classes out of blob.cc and into their own file.
This is a non-trivial amount of code and allows the compliex blob.cc
file to be a bit simpler.

There should be no behavior change.

Change-Id: I838484f2cd8acf3b33ba66085e4cb3834a92f88c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496764
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>