[fx] Rename fint_set feature to legacy_set
And invert its meaning. Whereas `fint_set` defaulted to false and made
`fx set` use the new implementation if true, `legacy_set` defaults to
true and `fx set` will use the new implementation if it's false.

The intent of the renaming is to make it more obvious and memorable, as
"fint" is not a widely known term. It should be safe because the new
feature has not yet been announced.

Bug: 68465
Change-Id: Ifec298cb8d36d9270830757156f4cb48d64be678
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497980
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Gary Boone <gboone@google.com>