[codec] Keep Magic used diagnosing fxbug.dev/67247
This Magic class is useful enough for debugging problems
that it seems worth keeping.  At the moment we don't
have any classes with a Magic<> member in them.  The
intent is to use a Magic<> member when we're trying to
narrow down anything that looks like corruption or
lack of type safety.

BUG: fxbug.dev/67247
Change-Id: I53c9f9eca7374f12c80dd8443f6a2346458a1268
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497000
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Reviewed-by: John Bauman <jbauman@google.com>